+++ tests/Js2IL.Test262.Tests/built-ins/Array/prototype/entries/ExecutionTests.cs
@@ -1,0 +1,14 @@
+using Js2IL.Test262.Tests.built_ins;
+
+
+namespace Js2IL.Test262.Tests.built_ins.Array.prototype.entries;
+
+
+public class ExecutionTests : DiskExecutionTestsBase
+{
+    public ExecutionTests() : base("built_ins.Array.prototype.entries") { }
+
+    [Fact(DisplayName = "iteration")]
+    public Task iteration()
+        => ExecutionTestFromFile("iteration");
+}